+++ packages/core/src/api/api/resources/datasetRunItems/client/requests/ListDatasetRunItemsRequest.ts
@@ -12,12 +12,8 @@
 export interface ListDatasetRunItemsRequest {
   datasetId: string;
   runName: string;
-  /**
-   * page number, starts at 1
-   */
+  /** page number, starts at 1 */
   page?: number;
-  /**
-   * limit of items per page
-   */
+  /** limit of items per page */
   limit?: number;
 }